feat: implement emulator tests.
This commit implements tests for the Azure Storage and AWS S3 backends to use
local storage emulators for tests.

Also fixes a panic due to a missing `break` in `cloud-copy`'s main function
that was caused by introducing a loop to drive the main future to completion
upon cancellation.

Note that Google Cloud Storage tests were not added as there currently isn't an
officially supported emulator.