feat: implement emulator tests. (#4)
* feat: implement emulator tests.

This commit implements tests for the Azure Storage and AWS S3 backends to use
local storage emulators for tests.

Also fixes a panic due to a missing `break` in `cloud-copy`'s main function
that was caused by introducing a loop to drive the main future to completion
upon cancellation.

Note that Google Cloud Storage tests were not added as there currently isn't an
officially supported emulator.

* fix: remove `macos` from test matrix until it supports Docker.

* chore: fix CI.

* chore: code review feedback.

Include how long the retry waited in the warning.

* chore: code review feedback.

* Removed the test environment variables in favor of proper configuration.
* The configuration now enables additional support for Azurite/Localstack and
  does not attempt to rewrite every URL (`az` and `s3` schemed URLs are
  rewritten when enabled, though).
* Updated the URLs in the tests to use Azurite/Localstack where appropriate.