feat: implement download caching. (#5)
* feat: implement download caching.

This commit implements downloading caching via the `http-cache-stream-reqwest`
library.

A `cache-dir` option was added to the CLI that enables the caching.

When enabled, HEAD and GET requests that are cachable are saved to the cache
directory. If the cache entry is a hit, the content is copied out of the cache
directory rather than contacting the upstream service.

Also implemented in this commit:

* The CLI now prints out statistics after the copy completes.

Also fixed in this commit:

* An incorrect `range` header caused an extra byte to be sent from the server;
  this didn't impact the copy as it does an exact read, but with the new
  caching support, the entire body needs to be read for it to end up in cache.
* The default parallelism ended up being a little too aggressive for high CPU
  count machines; dropping back to the previous default of the system's
  available parallelism.
* The user agent now has a link to the GitHub repository.
* Fixed progress updates not being sent at the correct intervals or with the
  correct number of bytes since the last event.

* chore: remove outdated doc comment.

* chore: update `http-cache-stream-reqwest` dependency version.

* chore: code review feedback.

Also treat middleware errors as transient.