feat: implement link to cached files. (#6)
* feat: implement link to cached files.

This commit implements creating hard links to cached files instead of making
copies.

If a hard link cannot be made, a copy is performed instead.

This commit also fixes the use case when `copy` is called to copy between local
file paths (supported, although not particularly useful). The copy was being
performed with a direct call to `fs::copy`, which did not emit any transfer
events. As a result, the statistics printed at the end of the run was always 0
files copied, 0 bytes transferred.

The fix was to wrap the source file stream with a transfer stream so that
transfer events are emitted and the statistics are then updated.

* chore: update CHANGELOG.

* fix: make transfer identifiers monotonic across all transfers.

* chore: refactor `cli` feature code.

This commit moves the `cli` feature code into the `cli` module.

Additionally, makes the `DisplayTimeDelta` utility type from `main.rs` as an
extension method on `TimeDelta` so that other CLI implementations can use it.

* Update src/cli.rs

Co-authored-by: Clay McLeod <3411613+claymcleod@users.noreply.github.com>

---------

Co-authored-by: Clay McLeod <3411613+claymcleod@users.noreply.github.com>