fix: fix empty uploads on Azure.
Apparently, `azurite` supports empty block uploads but Azure storage itself
does not; it gives a 400 on a content length of 0.

This commit fixes that by sending an empty block list for the blob if the file
size is zero.

Also makes the final `TransferStream` progress event more consistent by moving
it into a `Drop` implementation.