feat: support Windows.
This commit implements the necessary `fill_buffer` function which uses the
`windows` crate to do an overlapped `ReadFile` call and the requisite offset.

Also fixes a build error in the `Ctrl-C` handler and adds Windows and macOS to
build in CI.

Testing is still limited to Linux due to reliance on the Azurite and Localstack
containers.

Manually tested on Windows to round trip files to the three supported cloud
services.