feat: respect the `overwrite` option for remote URLs. (#11)
* feat: respect the `overwrite` option for remote URLs.

This commit implements checking to see if remote URLs exist to respect the
`overwrite` configuration option.

Previously, this option only checked for local destination existence.

* chore: update CHANGELOG.

* fix: fix empty uploads on Azure.

Apparently, `azurite` supports empty block uploads but Azure storage itself
does not; it gives a 400 on a content length of 0.

This commit fixes that by sending an empty block list for the blob if the file
size is zero.

Also makes the final `TransferStream` progress event more consistent by moving
it into a `Drop` implementation.

* chore: update CHANGELOG.

* chore: code review feedback.

* fix: just use a HEAD for checking if objects exist.

This commit removes the `exists` method that was added to the trait in favor of
just making a `HEAD` request upon the creation of an upload.