fix: remove OS-specific "read at offset" implementation.
This commit removes the OS-specific "read at offset" implementations in favor
of simply opening the file and seeking to the offset before a normal read; it
means we'll open the source file as many times there are blocks in the file
rather than just once, but the overhead of doing so is negligible.

Given that the maximum number of blocks on the supported storage backends are
10,000 (S3 and GCS) and 50,000 (Azure), the syscall overhead is unlikely to be
noticed by users especially when dwarfed by the network transfer times.