feat: support Windows. (#12)
* feat: support Windows.

This commit implements the necessary `fill_buffer` function which uses the
`windows` crate to do an overlapped `ReadFile` call and the requisite offset.

Also fixes a build error in the `Ctrl-C` handler and adds Windows and macOS to
build in CI.

Testing is still limited to Linux due to reliance on the Azurite and Localstack
containers.

Manually tested on Windows to round trip files to the three supported cloud
services.

* chore: update CHANGELOG.

* chore: code review feedback.

* chore: fix fmt.

* fix: remove OS-specific "read at offset" implementation.

This commit removes the OS-specific "read at offset" implementations in favor
of simply opening the file and seeking to the offset before a normal read; it
means we'll open the source file as many times there are blocks in the file
rather than just once, but the overhead of doing so is negligible.

Given that the maximum number of blocks on the supported storage backends are
10,000 (S3 and GCS) and 50,000 (Azure), the syscall overhead is unlikely to be
noticed by users especially when dwarfed by the network transfer times.

* Update src/transfer.rs

Co-authored-by: Adam C. Foltzer <acfoltzer@acfoltzer.net>

---------

Co-authored-by: Adam C. Foltzer <acfoltzer@acfoltzer.net>