fix: fix progress bar not being removed for failed transfers. (#14)
* fix: fix progress bar not being removed for failed transfers.

This commit fixes the `handle_events` function that is implemented as part of
the `cli` feature.

An incorrect order of operations when responding to the `TransferCompleted`
event resulted in not removing a transfer from the state if it failed; this
resulted in the progress bar continuing to be displayed for the transfer.

* chore: update CHANGELOG.