feat: store content digest metadata for uploaded objects.
This commit adds calculating SHA-256 content digests for uploaded objects,
using a metadata header to store a `Content-Digest` header value.

It also adds `get_content_digest` to retrieve the content digest of a given
URL, which issues a `HEAD` request and returns either a `Content-Digest` header
value (directly or via the metadata header) or a strong `ETag` header.

Adds the `--hash-algorithm` option to `cloud-copy` to specifying either `none`,
`sha256`, or `blake3` for calculating content digests.