feat: implement `size` and `exists` functions. (#30)
* feat: implement `size` and `exists` functions.

This commit implements a `size` function that makes a HEAD request and parses
the resulting content-length header, if there is one.

It also implements an `exists` function that determines if the given URL
"exists" by making a HEAD request. If the request is successful, we consider
the URL to "exist". Otherwise, if it is a cloud service URL, we check to see if
a walk of the URL would produce at least one result and consider it to "exist"
if so.

* chore: update CHANGELOG.